Created handling for customers and invoices
When clicking the buy button, it wil lredirect you to create a new customer. Once the new customer is created, it will redirect you to the invoice tab which will have info already filled out, only thing employee needs to do is determine what the status of the order is.